requirements.txt
fastapi
httpx
pydantic
pyinstaller
python-dotenv
starlette
uvicorn
